developer‘s API （not complete yet）
