+++ .gitignore
@@ -33,3 +33,6 @@
 *.exe
 *.out
 *.app
+
+# debug folder
+debug/
