+++ src/python/services/utils.py
@@ -1,5 +1,5 @@
 import numpy as np
-import h5py
+
 import tensorflow as tf
 import json
 import logging
@@ -7,145 +7,10 @@ import logging
 from dxl.learn.core import ThisHost
 
 
-logging.basicConfig(
-    format='[%(levelname)s] %(asctime)s [%(filename)s:%(lineno)d] %(message)s',
-    datefmt='%a, %d %b %Y %H:%M:%S',
-)
-logger = logging.getLogger('dxl.learn.graph.reconstruction')
-
-
-class ImageInfo:
-    def __init__(self, grid, center, size):
-        self.grid = grid
-        self.center = center
-        self.size = size
-
-
-class MapInfo:
-    def __init__(self, map_file):
-        self._map_file = map_file
-
-    def _maybe_broadcast_value(self,
-                               value,
-                               task_index=None,
-                               valid_type=(list, tuple)):
-        if task_index is None:
-            task_index = ThisHost
-        if isinstance(value, valid_type):
-            return value[task_index]
-        else:
-            return value
-
-    def map_file(self, task_index=None):
-        if task_index is None:
-            task_index = ThisHost.host().task_index
-        if isinstance(self._map_file, str):
-            return self._map_file
-        else:
-            return self._map_file[task_index]
-
-    def __str__(self):
-        result = {}
-        result['map_file'] = self.map_file()
-        return json.dumps(result)
-
-class DataInfo:
-    def __init__(self,
-                 lor_files,
-                 lor_shapes,
-                 lor_ranges=None,
-                 lor_step=None):
-        self._lor_files = lor_files
-        self._lor_shapes = lor_shapes
-        self._lor_ranges = lor_ranges
-        self._lor_step = lor_step
-
-    def _maybe_broadcast_value(self,
-                               value,
-                               task_index=None,
-                               valid_type=(list, tuple)):
-        if task_index is None:
-            task_index = ThisHost
-        if isinstance(value, valid_type):
-            return value[task_index]
-        else:
-            return value
-
-    def lor_file(self, axis, task_index=None):
-        if task_index is None:
-            task_index = ThisHost().host().task_index
-        if isinstance(self._lor_files[axis], str):
-            return self._lor_files[axis]
-        else:
-            return self._lor_files[axis][task_index]
-
-    def lor_range(self, axis, task_index=None):
-        if task_index is None:
-            task_index = ThisHost.host().task_index
-        if self._lor_ranges is not None:
-            return self._maybe_broadcast_value(self._lor_ranges[axis], task_index)
-        elif self._lor_step is not None:
-            step = self._maybe_broadcast_value(
-                self._lor_step[axis], task_index)
-            return [task_index * step, (task_index + 1) * step]
-        else:
-            return None
-
-    def lor_shape(self, axis, task_index=None):
-        if task_index is None:
-            task_index = ThisHost().host().task_index
-        if isinstance(self._lor_shapes[axis], (list, tuple)):
-            return self._lor_shapes[axis]
-        else:
-            return self._lor_shapes[axis][task_index]
-
-    def __str__(self):
-        result = {}
-        axis = ['x', 'y', 'z']
-        result['lor_file'] = {a: self.lor_file(a) for a in axis}
-        result['lor_range'] = {a: self.lor_range(a) for a in axis}
-        result['lor_shape'] = {a: self.lor_shape(a) for a in axis}
-        return json.dumps(result, indent=3, separators=(',', ': '))
-
-
-def load_data(file_name, lor_range=None):
-    if file_name.endswith('.npy'):
-        data = np.load(file_name)
-        if lor_range is not None:
-            data = data[lor_range[0]:lor_range[1], :]
-    elif file_name.endswith('.h5'):
-        with h5py.File(file_name, 'r') as fin:
-            if lor_range is not None:
-                data = np.array(fin['data'][lor_range[0]:lor_range[1], :])
-            else:
-                data = np.array(fin['data'])
-    return data
-
-
-# Load datas
-def load_local_data(data_info: DataInfo, task_index):
-
-    lors = {}
-    for a in ['x', 'y', 'z']:
-        msg = "Loading {} LORs from file: {}, with range: {}..."
-        logger.info(msg.format(
-            a, data_info.lor_file(a), data_info.lor_range(a)))
-        lors[a] = load_data(data_info.lor_file(a), data_info.lor_range(a))
-    logger.info('Loading local data done.')
-    return lors
-
-
-def load_local_effmap(map_info: MapInfo, task_index):
-    logger.info("Loading efficiency map from file: {}...".format(
-        map_info.map_file()))
-    emap = load_data(map_info.map_file())
-    return emap
-
-
-def ensure_float32(x):
-    if isinstance(x, np.ndarray) and x.dtype == np.float64:
-        return np.array(x, dtype=np.float32)
-    return x
+# def ensure_float32(x):
+#     if isinstance(x, np.ndarray) and x.dtype == np.float64:
+#         return np.array(x, dtype=np.float32)
+#     return x
 
 
 # def constant_tensor(x, name, ginfo):
@@ -165,44 +30,10 @@ def print_tensor(t, name=None):
         name, t.data, t.run()))
 
 
-def debug_tensor(t, msg):
+def debug_tensor(t, msg, logger):
     logger.debug("Debug {}, tensor: {}, (.data: {}):\n{}".format(
         msg, t, t.data, t.run()))
 
 
 def print_info(*msg):
     print('INFO', *msg)
-
-
-sample_reconstruction_config = {
-    'grid': [150, 150, 150],
-    'center': [0., 0., 0.],
-    'size': [150., 150., 150.],
-    'map_file': './debug/map.npy',
-    'x_lor_files': './debug/xlors.npy',
-    'y_lor_files': './debug/ylors.npy',
-    'z_lor_files': './debug/zlors.npy',
-    'x_lor_shapes': [100, 6],
-    'y_lor_shapes': [200, 6],
-    'z_lor_shapes': [300, 6],
-    'lor_ranges': None,
-    'lor_steps': None,
-}
-
-
-def load_reconstruction_configs(config=None):
-    if config is None:
-        c = sample_reconstruction_config
-    elif isinstance(config, str):
-        with open(config, 'r') as fin:
-            c = json.load(fin)
-    else:
-        c = config
-    image_info = ImageInfo(c['grid'], c['center'], c['size'])
-    map_info = MapInfo(c['map_file'])
-    data_info = DataInfo(
-        {a: c['{}_lor_files'.format(a)]
-         for a in ['x', 'y', 'z']},
-        {a: c['{}_lor_shapes'.format(a)]
-         for a in ['x', 'y', 'z']}, c['lor_ranges'], c['lor_steps'])
-    return image_info, map_info, data_info
